+++ frf/tests/test_serializers.py
@@ -440,6 +440,7 @@ class FakeProjectTestCase(BaseTestCase):
 
         json = res.json
         self.assertEqual(1, len(json))
+        print(json)
         company = json[0]
 
         self.assertEqual(3, len(company))
@@ -456,6 +457,24 @@ class FakeProjectTestCase(BaseTestCase):
 
         self.assertEquals(company['authors'][1]['uuid1'], str(self.ross.uuid1))
         self.assertEquals(company['authors'][1]['uuid2'], str(self.ross.uuid2))
+
+    def test_index_companies_v2(self):
+        res = self.simulate_get('/api/v2/companies/')
+        self.assertEqual(falcon.HTTP_200, res.status)
+
+        json = res.json
+        self.assertEqual(1, len(json))
+
+        company = json[0]
+        self.assertEqual(3, len(company))
+
+        self.assertIsInstance(company['authors'], list)
+        self.assertIsInstance(company['id'], int)
+        self.assertIsInstance(company['name'], str)
+
+        self.assertEquals(2, len(company['authors']))
+        self.assertEquals(self.adam.name, company['authors'][0])
+        self.assertEquals(self.ross.name, company['authors'][1])
 
     def test_update_company(self):
         c = models.Company(name='SOC')
